requirements.txt
google-cloud-storage==2.14.0
google-cloud-bigquery==3.17.0
python-dotenv==1.0.0
pendulum==2.1.2